Edit function enabled, works partially
I enabled the edit function, which works, but changes the wrong row. It seems like what's happening is that it replaces the xth column of the ordered database, not the xth column of the database as you get it when you run loadData(). Not sure how to fix this now, but will look into it.