fixed the dit function
I fixed the edit function. gs_edit_cells is really slow, so we should consider using a different function in the future, but since we're not sure if we will stick to googlesheets either, I guess we can leave it at that for now. Also the indexing is bit ugly, I added a constant 1 to get to the right row, has something to do with the header of the document, which is sometimes counted as a row and sometimes not. Need to look into that more.