Small pop-up making people aware of the new field to enter a new keyword
I added a small pop-up window that will be shown when people choose to enter a new keyword. I thought that in the previous version it was not very obvious that this new field had been added to the form. Let me know what you think. Originally I wanted to have the new field ("Add your new keyword") to be printed in red, but I couldn't get that to work...