Added a line that checks for double entries and brings a warning message if someone enters a last name that already exists
The only thing is that the data is still stored. So we would need to add that it is also deleted again, which I couldn't get to work now. Not sure if we want this check for double entires anymore anyway, though. It's easy to take out again, if we decide against it.